+++ frontend/src/javascripts/components/atoms/buttons/form-submit-btn.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const FormSubmitBtn = ({ label }) => (
+  <button type="submit" className="submit-btn">
+    {label}
+  </button>
+)
+
+export default FormSubmitBtn
