+++ frontend/src/javascripts/components/atoms/confirm-btn.js
@@ -2,8 +2,8 @@ import React from 'react'
 
 import '../../../stylesheets/confirm_btn.scss'
 
-const ConfirmBtn = message => {
-  return <button className="confirm-btn">{message.message}</button>
-}
+const ConfirmBtn = message => (
+  <button className="confirm-btn">{message.message}</button>
+)
 
 export default ConfirmBtn
