+++ frontend/src/javascripts/components/atoms/toppage-btn.js
@@ -1,16 +1,14 @@
 import React from 'react'
 import { TopPageImgs } from '../../constants/images'
 
-const TopPageButton = isSignIn => {
-  return (
-    <a className="toppage-btn">
-      {isSignIn ? (
-        <img className="toppage-btn-img" src={TopPageImgs['signUp']} />
-      ) : (
-        <img className="toppage-btn-img" src={TopPageImgs['signIn']} />
-      )}
-    </a>
-  )
-}
+const TopPageButton = isSignIn => (
+  <a className="toppage-btn">
+    {isSignIn ? (
+      <img className="toppage-btn-img" src={TopPageImgs['signUp']} />
+    ) : (
+      <img className="toppage-btn-img" src={TopPageImgs['signIn']} />
+    )}
+  </a>
+)
 
 export default TopPageButton
